// src/components/Dashboard/Job/JobDataType.d.ts
export interface JobDataType {
    companyName: string;
    companyId: string;
    jobTitle: string;
    internalJobTitle: string;
    primaryRecruiter: string;
    collaborator: string;
    collaboratorName: string;
    jobCategory: string;
    pipelineStatus: string;
    priority: string;
    jobType: string;
    jobHours: string;
    publicDescription: string;
    originalDescription: string;
    referenceNo: string;
    openings: string;
    startDate: Date;
    endDate: Date;
    businessUnit: string;
    remoteJob: string;
    remote_value: string;
    streetAddress: string;
    jobCity: string;
    stateOrPro: string;
    jobPostalCode: string;
    workType: string;
    specificLocation: string;
    candidateLocation: string;
    jobShift: any[];
    hoursWeek: string;
    payHours: string;
    payrateMin: string;
    payrateMax: string;
    payrateType: string;
    billRateMin: string;
    billRateMax: string;
    billRateType: string;
    dsBillMin: string;
    dsBillMax: string;
    dsBillType: string;
    firstName: string;
    lastName: string;
    email: string;
    phone: string;
    workflowId: string;
    workflowName: string;
    formId: string;
    formName: string;
    jobId: string;
    json: string;
    workflowDetails: WorkflowDetails;
}

interface WorkflowDetails {
    workflow_edit: boolean;
    isdelete: boolean;
    ispause: boolean;
    workflow_job_id: number;
    workflowid: number;
    workflowname: string;
}